feat(camera.zig): add support for anti-aliasing by averaging multiple samples per pixel refactor(image.zig): modify write_color function to scale color by the number of samples refactor(interval.zig, rtweekend.zig): change data type from f64 to f32 for better performance feat(main.zig): add samples_per_pixel parameter to Camera.init function call for anti-aliasing refactor(rtweekend.zig): change Infinity constant from f64 to f32 for consistency with other changes

+++ src/interval.zig
@@ -7,12 +7,12 @@ const rtweekend = @import("rtweekend.zig");
 const Infinity = rtweekend.Infinity;
 
 pub const Interval = struct {
-    min: f64,
-    max: f64,
+    min: f32,
+    max: f32,
 
     const Self = @This();
 
-    pub fn init(min: f64, max: f64) Self {
+    pub fn init(min: f32, max: f32) Self {
         return .{ .min = min, .max = max };
     }
 
@@ -20,15 +20,15 @@ pub const Interval = struct {
         return .{ .min = -Infinity, .max = Infinity };
     }
 
-    pub fn contains(self: *Self, value: f64) bool {
+    pub fn contains(self: *Self, value: f32) bool {
         return self.min <= value and value <= self.max;
     }
 
-    pub fn surrounds(self: *Self, value: f64) bool {
+    pub fn surrounds(self: *Self, value: f32) bool {
         return self.min < value and value < self.max;
     }
 
-    pub fn clamp(self: *Self, value: f64) f64 {
+    pub fn clamp(self: *Self, value: f32) f32 {
         return math.clamp(value, self.min, self.max);
     }
 };
